layout/style/test/test_transitions_with_disabled_properties.html

<!doctype html>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1265611
-->
<head>
  <meta charset=utf-8>
  <title>Test for bug 1265611</title>
  <script type="application/javascript"
    src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <style>
  #display {
    transition: all 0.01s;
  }
  </style>
</head>
<body>
<a target="_blank"
  href="https://bugzilla.mozilla.org/show_bug.cgi?id=1265611">Mozilla Bug
  1265611</a>

<div id="display"></div>
<pre id="test">
<script>
SimpleTest.waitForExplicitFinish();

/*
 * This tests for transitions generated on the -webkit-text-fill-color property.
 * This property has an initial value of 'currentcolor' so by triggering a
 * transition on the 'color' property we also--at least at the point when
 * this test was written--trigger a transition on the -webkit-text-fill-color
 * property (that behavior may change in bug 1260543).
 *
 * However, before beginning the test we disable -webkit-text-fill-color by
 * setting layout.css.prefixes.webkit to false. This code tests that we don't
 * end up triggering a transition on the (disabled) property in that case.
 */
SpecialPowers.pushPrefEnv({'set': [['layout.css.prefixes.webkit', false],
                                   ['dom.animations-api.core.enabled', true]] },
  () => {
    var display = document.getElementById('display');
    display.style.color = 'green';

    var transitionedProperties =
      display.getAnimations().map(transition => transition.transitionProperty);

    ok(!transitionedProperties.includes('-webkit-text-fill-color'),
       'We should not fire transitions for properties disabled by prefs');
    SimpleTest.finish();
  }
);
</script>
</pre>
</body>
</html>
